Fix coloring of multi-word options.
If you have a multi-word option such as:

 has foo_bar => (is => 'ro', metaclass => 'Getopt', cmd_flag => 'get-opt');

Then this is handled with the getopt options `--foo-bar`

The colorized output only colors `--foo` and the `-bar` is uncolored
because the formatter stops at the first non-word character.

Given that `-` is a common word separator for multi-word options, it
makes sense to include this in the list of chars that are colored.